Generate a derivative file to avoid license issues
I do not know if there actually is a license that prohibits me from directly using the database. Using a derivative file is however the safer option until I am definitely sure that there are no issues with using the original database. We don't want to get anyone into trouble now, do we?