Move redundancy in styles
Because the formatting in the lookup window is broken regardless of using a style section, we can use inline styles for both manual and automatic lookups. As a result it is also easier to edit the mappings between the styles (the "styles" dict at the top of the nhk_pronunciation.py).