Add bulk-add option for pronunciation
A button has been added to the Browser menu that will allow you to permanently write the pronunciation to the selected notes.

The increase in database size was relatively limited (in my case 14.3MB -> 15.6MB).

By default the bulk-add option will not overwrite fields, but you may set the "regenerate_readings" to True to change this behavior.