Fix undefined variable 'src' (or 'dst') error
The focusLost routine (called when adding cards) had its own lookup function for gettin the source and destination field names. It did however not properly initialize the src/dst variables to None. This resulted in an undefined variable error if the model name included 'japanese' (case-insensitive), but field names of that model did not match any of the srcField or dstField entries.