Support display of pronunciation in hiragana
A flag in the nhk_pronunciation.py file allows the user to use hiragana pronunciations instead of the default katakana pronunciations. The convergence is done on-the-fly, and does not affect the pickle or underlying database.

When regenerating the pronunciation using the Bulk-Add feature, do remember that another flag might prevent the add-on from overwriting a non-emmpty (katakana-containing) field. To overwrite all existing katakana entries, one would therefore also have to (temporarily) enable this option.

When deemed stable enough, and enough users use this feature, consider moving it to a UI setting (instead of hardcoded in the source).

Closes #2.