Upgrade to support Anki 2.1
The user config section in the Python file has been moved to a config.json
file, such that it can be edited from the Anki user interface (and such
that it persists over upgrades).

The default shortcut Ctrl+6 to add pronunciation has been removed, as it
typically conflicts with other shortcuts.

The dependency on the Japanese add-on was also removed, although it is not
yet known what effect this has. For example, "onLookupPronunciation" used
to call the initLookup() method of the Japanese add-on, but it seems to
work fine without.

Closes #6