Fixed issues w/working on multiple words
There are 2 main issues I've found with the pronunciation add-on, but I've found a way to fix most of them with some quick tweaks.

1) The card type name itself needs to have the word "japanese" somewhere in it. 2) It has to be one word, which often means it's without kana (especially without grammatical stuff attached like particles), so things like this won't work (commas separate examples): 後で, 川上・川下 or other fields with punctuation, whole sentences, things with readings attached (because of the weird characters), e.g.: 後れた【おくれた】.

I've fixed the requirement of a japanese card type, the unnecessary kana problem (for most use-cases), the punctuation/separators problem, and the problem of multiple words strung together. This fix obviously focuses on individual words rather than prosody (which you should use OJAD/suzuki-kun for). Since the cleanest quick solution I could think is based the solution off Japanese Support, this splits words according to the Japanese add-on (which uses Mecab), rather than simply checking the dictionary. To get around this problem, I made it so that will try a punctuation split (including Japanese pronunciation) by default. To get the Mecab-type sentence-splitting (which will correctly look up words even if they are separated by kana), you have to change srcFields to 'Sentence' - but if someone can update this functionality to work automatically if you use a Sentence field (perhaps even if you have both Sentence and Expression, e.g. adding different pronunciations to different destinations), that'd be great. Even better, perhaps, might be some automatic detection of whether something is a sentence or just a string of words, but that might be tricky.