Fully automated multi-word functionality
Now you don't need to worry about whether something is a sentence or a series of words - it'll just work properly.

First, it checks if it can match every word using a simple word-splitting based on Japanese/English punctuation and accounting for possible 送り仮名 or just random kana that needn't be there. If it can't find a match for every entry, it switches to the more liberal sentence-level splitter (the assumption being that either these words are too complicated and thus have no entry, but they might be broken down further by sentence-level analysis; or that these are actually multiple words separated by kana, or compounded to remove instances of 'no' or something).