fixed kana stripping
It wasn't properly parsing some things that normally have kana at the end, but was instead stripping the kana first. Fixed that problem. It still will only work if the entry is in the dictionary. For example, くやしい doesn't work, but 悔しい does. Since all entries have a katakana pronunciation, a work-around for that problem would be to use the katakana entries for all-kana words that aren't found initially, but that will cause edge-case problems for homographs, so explicitly notifying the user that it's just a guess would be best in those cases.