Bring back Python 2 / Anki 2.0 support
Note that we are explicit about importing/using io's open(), such that
the behavior is the same between Python 2 and 3.

The configuration of the Anki 2.0 add-on is now also using the a JSON
config file. Anki 2.0 does not support editing this config from the UI,
but a regular file editor will do.

The realpath to normpath change is needed to avoid resolving symlinks.
This is typically only useful for developers.