Support automatic lookup on delimited expressions
If the expression is a delimited one with a common separator like a slash
(or the Japanese equivalent), we try a lookup on each sub-expression if we
cannot find one for the overall expression.

Note that this does not support lookup up pronunciations for arbitrary
sentences or conjugations. That will need something like mecab to do the
splitting for us.