Support splitting expression with Mecab
If Mecab is available we can split the an expression for which no
pronunciations are found to see if sub-expression do have entries in our
pronunciation database.

The lookup fallback using Mecab is enabled by default, and can be disabled
by setting "useMecab" to False in config.json.

Note that the Japanese add-on is required for this functionality to work.
If useMecab is true, but Mecab cannot be found (because the Japanese
add-on is missing), an error message is shown and further lookup with
Mecab is disabled for the session.